Rename components page
